+++ One-Umbrella.DAL/Interfaces/IRatingRepository.cs
@@ -10,7 +10,7 @@ namespace OneUmbrella.DAL.Interfaces
     public interface IRatingRepository
     {
         IEnumerable<Rating> getAllByRestaurant(int restaurantId, bool isHuman);
-        int countForOneRestaurant(int restaurantId);
+        decimal countForOneRestaurant(int restaurantId);
         bool create(Rating rating);
         bool update(Rating rating);
         bool delete(int humanId, int restaurantId);
