+++ One-Umbrella.Domain/Entities/Restaurant.cs
@@ -15,6 +15,6 @@ namespace OneUmbrella.Domain.Entities
         public string RestaurantCity { get; set; }
         public string RestaurantPostCode { get; set; }
         public string RestaurantDescription { get; set; }
-        public int RestaurantRating { get; set; }
+        public decimal RestaurantRating { get; set; }
     }
 }
